// projects/step-frontend/src/lib/modules/artefacts/types/operator-type.enum.ts
export enum OperatorType {
  EQUALS = 'EQUALS',
  BEGINS_WITH = 'BEGINS_WITH',
  CONTAINS = 'CONTAINS',
  ENDS_WITH = 'ENDS_WITH',
  MATCHES = 'MATCHES',
  LESS_THAN = 'LESS_THAN',
  LESS_THAN_OR_EQUALS = 'LESS_THAN_OR_EQUALS',
  GREATER_THAN = 'GREATER_THAN',
  GREATER_THAN_OR_EQUALS = 'GREATER_THAN_OR_EQUALS',
  IS_NULL = 'IS_NULL',
}